batch: Move replacement units into normalizer
The ownership package now separates baseline references and presence claims
from the aggregate ownership model, leaving replacement-unit normalization
as the last internal cycle through that aggregate.

ReplacementUnit and ReplacementUnitOrigin still sit on ownership.py while
normalization, remapping, rebuild, and translator callers treat those
records as replacement-unit metadata. That forces the normalizer to import
the aggregate and keeps the batch ownership graph cyclic.

This commit moves the replacement-unit records into
batch.ownership_replacement_units and updates production callers to import
records from the same module as normalization.

Subsequent commits will update tests and pin the architecture guard so the
batch ownership package remains acyclic.
