+++ src/git_stage_batch/batch/ownership_detachment.py
@@ -4,9 +4,10 @@ from __future__ import annotations
 
 from ..core.buffer import LineBuffer
 from .absence_content import copy_absence_content as _copy_absence_content
-from .ownership import AbsenceClaim, BatchOwnership, ReplacementUnit
+from .ownership import AbsenceClaim, BatchOwnership
 from .ownership_acquisition import AcquiredBatchOwnership
 from .ownership_claims import PresenceClaim
+from .ownership_replacement_units import ReplacementUnit
 
 
 def acquire_detached_batch_ownership(
