+++ src/git_stage_batch/batch/ownership_unit_types.py
@@ -6,7 +6,8 @@ from dataclasses import dataclass
 from enum import Enum
 
 from ..core.line_selection import LineRanges
-from .ownership import AbsenceClaim, ReplacementUnitOrigin
+from .ownership import AbsenceClaim
+from .ownership_replacement_units import ReplacementUnitOrigin
 
 
 class OwnershipUnitKind(Enum):
